NIP-Igor

+++ src/app/view/admin/local.component.ts
@@ -48,7 +48,11 @@ export class LocalComponent implements OnInit {
   inserir(): void {
     this.localService.addLocal(this.local)
       .subscribe(response => {
-        if (response.Status == 0) {
+        if (response === undefined){
+          this.mostraErro('Não foi possível realizar o cadastro do usuário.')
+        }
+        else if (response.Status == 0) {
+          this.limpar()
           this.getLocais();
         }
         else {
@@ -61,11 +65,11 @@ export class LocalComponent implements OnInit {
     this.localService.updateLocal(this.local)
       .subscribe(response => {
         if (response.Status == 0) {
-          this.locais[this.selectedIndex].Nome = response.Local.Nome;
-          this.locais[this.selectedIndex].Reservavel = response.Local.Reservavel;
-          this.locais[this.selectedIndex].Disponivel = response.Local.Disponivel;
-          this.locais[this.selectedIndex].Tipo = response.Local.Tipo;
-
+          this.locais[this.selectedIndex].Nome = this.local.Nome;
+          this.locais[this.selectedIndex].Reservavel = this.local.Reservavel;
+          this.locais[this.selectedIndex].Disponivel = this.local.Disponivel;
+          this.locais[this.selectedIndex].Tipo = this.local.Tipo;
+          this.limpar()
         }
         else {
           this.mostraErro(response.Detalhes)
@@ -73,16 +77,23 @@ export class LocalComponent implements OnInit {
       });
   }
 
-  remover(id): void {
-    this.localService.deleteLocal(id)
+  remover(index): void {
+    if (confirm('Confirma remoção de ' + this.locais[index].Nome)) {
+      let id = this.locais[index].Id
+      this.localService.deleteLocal(id)
       .subscribe(response => {
-        if (response.Status == 0) {
+        if (response === undefined){
+          this.mostraErro('Não foi possível realizar a remoção do local.')
+        }
+        else if (response.Status == 0) {
+          this.limpar()
           this.getLocais();
         }
         else {
           this.mostraErro(response.Detalhes)
         }
       });
+    }
   }
 
   carregar(index): void {
@@ -96,6 +107,14 @@ export class LocalComponent implements OnInit {
   }
 
   registrar(): void {
+    if (this.local.Nome === undefined || this.local.Nome === ''){
+      this.mostraErro('Digite o nome do local.')
+      return
+    }
+    if (this.local.Tipo == -1){
+      this.mostraErro('Escolha um tipo de local.')
+      return
+    }
     if (this.local.Id === undefined) {
       this.inserir()
     }
